+++ src/assets/index.ts
@@ -1,3 +1,5 @@
 export * from './AssetsDeleteFacade';
 export * from './AssetsQueryFacade';
 export * from './PublicAssets';
+export * from './BucketCreator';
+export * from './AssetsArrayRef';